Spring security merged

+++ src/main/java/com/harsh/journalApp/service/JournalServiceEntry.java
@@ -3,13 +3,16 @@
 
 package com.harsh.journalApp.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.harsh.journalApp.entity.JournalEntry;
+import com.harsh.journalApp.entity.User;
 import com.harsh.journalApp.repository.JournalEntryRepository;
 
 @Component
@@ -17,10 +20,19 @@ public class JournalServiceEntry {
     
     @Autowired
     private JournalEntryRepository journalEntryRepository; 
+    @Autowired
+    private UserService userService;
 
+    @Transactional
+    public void saveEntry(JournalEntry journalEntry, String userName){
+        User  user =  userService.findByUserName(userName);
+        journalEntry.setDate(LocalDateTime.now());
+        JournalEntry saved = journalEntryRepository.save(journalEntry); 
+        user.getJournalEntries().add(saved);
+        userService.saveEntry(user);
+    }
     public void saveEntry(JournalEntry journalEntry){
-        
-        journalEntryRepository.save(journalEntry); 
+        journalEntryRepository.save(journalEntry);
     }
     public List<JournalEntry> getAll(){
 
@@ -30,9 +42,20 @@ public class JournalServiceEntry {
         
         return journalEntryRepository.findById(id).orElse(null);
     }
-    public void deleteById(ObjectId id){
-
-        journalEntryRepository.deleteById(id);
+    
+    @Transactional
+    public void deleteById(ObjectId id ,String userName){
+       try{
+        User  user =  userService.findByUserName(userName);
+        boolean removed = user.getJournalEntries().removeIf(x ->x.getId().equals(id));
+        if(removed)
+        {
+            userService.saveEntry(user);
+            journalEntryRepository.deleteById(id);
+        }
+       }catch(Exception e){
+        throw new RuntimeException("Error while deleting entry",e);
+       }
         
     }
